+++ dashboard/roles/server_processing.php
@@ -31,7 +31,8 @@ $primaryKey = 'id';
 $columns = array(
     array('db'  => 'id', 'dt'   => 0,
             'formatter' => function($d){
-                return "<input type='checkbox'  onchange='toggleCheckbox()' name='checkRole[]' value='$d'/>";
+                if($d != 1)
+                    return "<input type='checkbox'  onchange='toggleCheckbox()' name='checkRole[]' value='$d'/>";
             }
         ),
     array( 'db' => 'id', 'dt' => 1 ),
@@ -47,10 +48,15 @@ $columns = array(
         'db'        => 'id',
         'dt'        => 4,
         'formatter' => function ($d , $row){
-            return "
-            <a href='?do=Edit&id=$d' class='btn btn-warning'> <i class='fa fa-pen'></i></a>
-            <a onclick='confirmation(event)' href='?do=Delete&id=$d' class='btn btn-danger'> <i class='fa fa-trash'></i></a>
-            ";
+            if($d != 1){
+                return "
+                <a onclick='confirmation(event)' href='?do=Delete&id=$d' class='btn btn-danger'> <i class='fa fa-trash'></i></a>
+                ";
+            }
+            else{
+                return "<a href='?do=Edit&id=$d' class='btn btn-warning'> <i class='fa fa-pen'></i></a>";
+            }
+            
         }
     )
